+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 
 ## [unreleased]
 
+ * Portal: The *MashroomPortalStateService* prefixes now data in the browser storage to avoid conflicts.
+   So, its save now to use it to persist the application state like this:
+   ```js
+        stateService.setLocalStoreStateProperty('state', store.getState());
+   ```
  * Portal: Made it possible to use environment variables in the plugin definition of remote portal apps as well
  * Portal: Call the *willBeRemoved* lifecycle hook of all apps on page unload; this gives the app a chance to do some
    cleanup or persist its state properly